Refactor JwtProfile and OidcProfile
