+++ pac4j-jwt/src/main/java/org/pac4j/jwt/profile/JwtProfile.java
@@ -1,10 +1,7 @@
 package org.pac4j.jwt.profile;
 
-import org.pac4j.core.profile.CommonProfile;
-import org.pac4j.jwt.JwtClaims;
-
-import java.util.Date;
-import java.util.List;
+import org.pac4j.core.profile.jwt.AbstractJwtProfile;
+import org.pac4j.core.profile.jwt.JwtClaims;
 
 /**
  * Represents a user profile based on a JWT.
@@ -12,33 +9,9 @@ import java.util.List;
  * @author Misagh Moayyed
  * @since 1.8.2
  */
-public class JwtProfile extends CommonProfile {
+public class JwtProfile extends AbstractJwtProfile {
 
     private static final long serialVersionUID = -1688563185891542198L;
-
-    public String getSubject() {
-        return getId();
-    }
-
-    public String getIssuer() {
-        return (String) getAttribute(JwtClaims.ISSUER);
-    }
-
-    public List<String> getAudience() {
-        return retrieveAudience();
-    }
-
-    public Date getExpirationDate() {
-        return (Date) getAttribute(JwtClaims.EXPIRATION_TIME);
-    }
-
-    public Date getNotBefore() {
-        return (Date) getAttribute(JwtClaims.NOT_BEFORE);
-    }
-
-    public Date getIssuedAt() {
-        return (Date) getAttribute(JwtClaims.ISSUED_AT);
-    }
 
     public String getJwtId() {
         return (String) getAttribute(JwtClaims.JWT_ID);
