+++ pac4j-oidc/src/main/java/org/pac4j/oidc/profile/OidcProfileDefinition.java
@@ -5,6 +5,7 @@ import com.nimbusds.oauth2.sdk.token.BearerAccessToken;
 import com.nimbusds.oauth2.sdk.token.RefreshToken;
 import org.pac4j.core.profile.converter.Converters;
 import org.pac4j.core.profile.definition.CommonProfileDefinition;
+import org.pac4j.core.profile.jwt.JwtClaims;
 import org.pac4j.oidc.profile.converter.OidcLongTimeConverter;
 
 import java.util.Arrays;
@@ -37,16 +38,11 @@ public class OidcProfileDefinition<P extends OidcProfile> extends CommonProfileD
     public static final String ACCESS_TOKEN = "access_token";
     public static final String ID_TOKEN = "id_token";
     public static final String REFRESH_TOKEN = "refresh_token";
-    public static final String ISSUER          = "iss";
-    public static final String SUBJECT         = "sub";
-    public static final String EXPIRATION_TIME = "exp";
-    public static final String ISSUED_AT       = "iat";
     public static final String AUTH_TIME       = "auth_time";
     public static final String NONCE           = "nonce";
     public static final String ACR             = "acr";
     public static final String AMR             = "amr";
     public static final String AZP             = "azp";
-    public static final String NBF             = "nbf";
 
     public OidcProfileDefinition() {
         super(x -> (P) new OidcProfile());
@@ -74,8 +70,8 @@ public class OidcProfileDefinition<P extends OidcProfile> extends CommonProfileD
             return null;
         });
         // TODO: birthdate, address
-        Arrays.stream(new String[] {SUBJECT, ISSUER, NONCE, ACR, AZP}).forEach(a -> primary(a, Converters.STRING));
-        Arrays.stream(new String[] {EXPIRATION_TIME, ISSUED_AT, NBF}).forEach(a -> primary(a, Converters.DATE_TZ_GENERAL));
+        Arrays.stream(new String[] {JwtClaims.SUBJECT, JwtClaims.ISSUER, NONCE, ACR, AZP}).forEach(a -> primary(a, Converters.STRING));
+        Arrays.stream(new String[] {JwtClaims.EXPIRATION_TIME, JwtClaims.ISSUED_AT, JwtClaims.NOT_BEFORE}).forEach(a -> primary(a, Converters.DATE_TZ_GENERAL));
         primary(AUTH_TIME, new OidcLongTimeConverter());
     }
